# Optimize queries on histogram statistics