--- a/logicalplan/histogram_stats.go
+++ b/logicalplan/histogram_stats.go
@@ -1,3 +1,6 @@
+// Copyright (c) The Thanos Community Authors.
+// Licensed under the Apache License 2.0.
+
 package logicalplan
 
 import (