--- a/logicalplan/histogram_stats.go
+++ b/logicalplan/histogram_stats.go
@@ -12,11 +12,26 @@ import (
 type DetectHistogramStatsOptimizer struct{}
 
 func (d DetectHistogramStatsOptimizer) Optimize(plan Node, _ *query.Options) (Node, annotations.Annotations) {
-	var decodeStats bool
+	return d.optimize(plan)
+}
+
+func (d DetectHistogramStatsOptimizer) optimize(plan Node) (Node, annotations.Annotations) {
+	var (
+		stop        bool
+		decodeStats bool
+	)
 	Traverse(&plan, func(node *Node) {
+		if stop {
+			return
+		}
 		switch n := (*node).(type) {
 		case *VectorSelector:
 			n.DecodeNativeHistogramStats = decodeStats
+		case *Binary:
+			n.LHS, _ = d.optimize(n.LHS)
+			n.RHS, _ = d.optimize(n.RHS)
+			stop = true
+			return
 		case *FunctionCall:
 			if n.Func.Name == "histogram_count" || n.Func.Name == "histogram_sum" {
 				decodeStats = true