--- a/storage/prometheus/histograms_test.go
+++ b/storage/prometheus/histograms_test.go
@@ -1,3 +1,6 @@
+// Copyright (c) The Thanos Community Authors.
+// Licensed under the Apache License 2.0.
+
 package prometheus
 
 import (